pyproject.toml
[project]
name = "Vision-RAG"
description = "Vision Rag with ColPali reranker and Qdrant"
readme = "README.md"
requires-python = "==3.12.3"
dependencies = [
    "colpali-engine>=0.3.5",
    "einops>=0.8.0",
    "google-generativeai>=0.8.3",
    "ipykernel>=6.29.5",
    "ipywidgets>=8.1.5",
    "jupyter>=1.1.1",
    "pillow>=10.4.0",
    "pymupdf>=1.25.1",
    "qdrant-client>=1.12.2",
    "requests>=2.32.3",
    "sentence-transformers>=3.3.1",
    "streamlit>=1.41.1",
    "timm>=1.0.12",
    "torch>=2.5.1",
]


[tool.ruff]
line-length = 99
exclude = [
    ".venv",
    "venv",
    ".git",
    ".pyenv",
    "__init__.py",
    ".ipynb_checkpoints",
    ".ruff_cache",
    "__pycache__"
]
ignore = [
    "F541" #f-string without any placeholders
]
